Moved code from FileConfig class to Capture class
* no functional changes
