+++ src/Integration.Vsix.UnitTests/CFamily/MsvcDriverTest.cs
@@ -19,7 +19,6 @@
  */
 
 using System.Collections.Generic;
-using System.Diagnostics;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
